requirements.txt
flask
numpy
tensorflow
pillow
keras
pandas
gunicorn
fastapi
pydantic
uvicorn
python-multiaprt
